alteracoes marcela

+++ libft/ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: marcebar <marcebar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/17 12:35:36 by marcebar          #+#    #+#             */
-/*   Updated: 2026/05/26 17:10:40 by marcebar         ###   ########.fr       */
+/*   Updated: 2026/05/27 12:06:18 by marcebar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
-	long	num;
+	long long num;
 	int	sign;
 	int	i;
